+++ plan.md
@@ -8,3 +8,4 @@
 
 ✅ `src/shared/ui/Sidebar` を `src/widgets` へ移動する
 ✅ `src/shared/ui/Modal/PostModal` を `src/features/create-post/ui` へ移動する
+✅ `src/shared/ui/NewPostForm/NewPostForm.tsx` を `src/features/create-post/ui` へ移動する
