+++ src/features/create-post/index.ts
@@ -1,1 +1,2 @@
 export { PostModal } from "./ui/PostModal";
+export { NewPostForm } from "./ui/NewPostForm";
