+++ src/pages/top/ui/TopPageClient.tsx
@@ -1,8 +1,8 @@
 "use client";
 
+import { NewPostForm } from "@/features/create-post";
 import type { PostType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { NewPostForm } from "@/shared/ui";
 import { PostsList } from "@/widgets/posts-list";
 import { Sidebar } from "@/widgets/sidebar";
 import { Box, Flex } from "@mantine/core";
