+++ src/shared/ui/index.ts
@@ -1,5 +1,4 @@
 export { ImageUpload } from "./ImageUpload/ImageUpload";
-export { NewPostForm } from "./NewPostForm/NewPostForm";
 export { CenteredLoader } from "./CenteredLoader/CenteredLoader";
 export { ContentPart, parseContent } from "./ContentPart/ContentPart";
 export { DateFormat } from "./DateFormat/DateFormat";
